AggregatedNumbers/AggregatedCrawlerNumbers: -update to better support mirrors -fixes thread 67955
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@32852 ebf7c1c2-ba36-0410-9fe8-c592906822b4

Former-commit-id: 46b1ac59d7dfb849d6a0019441d442f024bb4bc3

+++ src/org/jdownloader/api/jd/AggregatedNumbersAPIStorable.java
@@ -46,11 +46,11 @@ public class AggregatedNumbersAPIStorable implements Storable {
         return aggregated.getEta();
     }
 
-    public Long getRunning() {
+    public Integer getRunning() {
         return aggregated.getRunning();
     }
 
-    public Long getConnections() {
+    public Integer getConnections() {
         return aggregated.getConnections();
     }
